requirements.txt
# Core Django Framework
Django==5.2.8
asgiref==3.11.0
sqlparse==0.5.3

# REST API Framework
djangorestframework==3.16.1

# CORS Support
django-cors-headers==4.9.0

# Authentication (Optional - for social login features)
# Uncomment if using social authentication features
# django-allauth==65.2.0
# django-cors-headers==4.9.0

# Database (Optional - for production databases)
# Uncomment when switching from SQLite to PostgreSQL
# psycopg2-binary==2.9.9
# Or for MySQL:
# mysqlclient==2.2.4

# Additional Development Tools (Optional)
# Uncomment for enhanced development experience
# django-debug-toolbar==4.4.6
# django-extensions==3.2.3
# pillow==10.4.0  # For image/file handling
